Update parse_urls.py
A rewrite of the discover plugin, using plugins as input and an imported instance of the html input plugin for parsing. Not entirely functional, as entries don't yet make it from the imported instance of html.py into task.entries.
More significantly, a function is needed to cache the results so that only new results need be parsed. With my own configuration, I average 150 entries every 10 minutes. Parsing each of them would be ill.