Firming up plugin, initial code for caching and fails.
Removed redundant discover code, and added the beginnings of caching and handling of failed entries. The cache code is still faulty, as it requires as the container needs to be iterable when empty. Failure handling will eventually intercept backlog entries and either update them to run on their own with the latest URLs or remove them altogether and inject them back into the normal entries.