Reduced scope of abort backlog and corrected reparse error exception. (See description.)
The scope of entries backlogged on task abort has been reduced, by excluding rejected entries in any phase except output, and only  including accepted entries then. The ability to determine the phase in which a task was aborted relies on "aborted_phase", a new value, added to and controlled by task.py. Corrected the exception of errors upon reparsing of any entries.